<Chore> :pull 받기 전 commit ---본문--- 페이먼츠 중간 수정 및 pull전 커밋 --- 꼬리말 ---  #11

+++ payments/models.py
@@ -12,7 +12,7 @@ class Payment(models.Model):
     업데이트 날짜 :
     '''
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    amount= models.PositiveIntegerField()
+    amount= models.DecimalField(max_digits=10, decimal_places=2)
     payment_Token = models.CharField(max_length=255)
     created_at = models.DateTimeField(auto_now_add=True)
